+++ app/models/inventory.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, Float, ForeignKey
+from sqlalchemy import Column, Integer, ForeignKey
 
 from app.db.base import Base
 
@@ -7,5 +7,5 @@ class Inventory(Base):
 
     sku_id = Column(Integer, ForeignKey("skus.id"), primary_key=True)
     warehouse_id = Column(Integer, ForeignKey("warehouses.id"), primary_key=True)
-    total_quantity = Column(Float, default=0, nullable=False)
-    reserved_quantity = Column(Float, default=0, nullable=False)
+    total_quantity = Column(Integer, default=0, nullable=False)
+    reserved_quantity = Column(Integer, default=0, nullable=False)
