+++ app/models/order_item.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, Float, ForeignKey
+from sqlalchemy import Column, Integer, Numeric, ForeignKey
 from sqlalchemy.orm import relationship
 
 from app.db.base import Base
@@ -9,9 +9,9 @@ class OrderItem(Base):
     id = Column(Integer, primary_key=True, index=True)
     order_id = Column(Integer, ForeignKey("orders.id"), nullable=False)
     sku_id = Column(Integer, ForeignKey("skus.id"), nullable=False)
-    quantity = Column(Float, nullable=False)
-    unit_price = Column(Float, nullable=False)
-    discount_amount = Column(Float, default=0, nullable=False)
+    quantity = Column(Integer, nullable=False)
+    unit_price = Column(Numeric(12, 2), nullable=False)
+    discount_amount = Column(Numeric(12, 2), default=0, nullable=False)
 
     order = relationship("Order", back_populates="items")
     taxes = relationship("OrderItemTax", back_populates="order_item", cascade="all, delete-orphan")
