+++ app/models/order_item_batch.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, Float, ForeignKey
+from sqlalchemy import Column, Integer, ForeignKey
 from sqlalchemy.orm import relationship
 
 from app.db.base import Base
@@ -9,7 +9,7 @@ class OrderItemBatch(Base):
     id = Column(Integer, primary_key=True, index=True)
     order_item_id = Column(Integer, ForeignKey("order_items.id"), nullable=False)
     batch_id = Column(Integer, ForeignKey("sku_batches.id"), nullable=False)
-    quantity = Column(Float, nullable=False)
+    quantity = Column(Integer, nullable=False)
 
     order_item = relationship("OrderItem", back_populates="order_item_batches")
     batch = relationship("SKUBatch", back_populates="order_item_batches")
